pyproject.toml
[project]
name = "okta-mcp-em-python"
version = "1.0.0"
description = "MCP Server for Okta Identity Governance - Entitlement Management"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
    {name = "Ashwin Ramnarayanan"}
]
keywords = ["okta", "mcp", "entitlements", "identity", "governance", "iga"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration :: Authentication/Directory",
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "boto3>=1.34.0",
    "pydantic>=2.0.0"
]

[project.urls]
Homepage = "https://github.com/ashwinrama/okta-mcp-em-python"
Documentation = "https://github.com/ashwinrama/okta-mcp-em-python#readme"
Repository = "https://github.com/ashwinrama/okta-mcp-em-python"
Issues = "https://github.com/ashwinrama/okta-mcp-em-python/issues"

[project.scripts]
okta-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools"]
